fix(docs): stop preview code from listing imports it never uses
the import filter matched identifiers as substrings, so a specimen that renders BreadcrumbSeparator also carried an unused Separator import. eight design specimens print at least one import their snippet cannot use, and Table prints the whole Dialog group. match on word boundaries instead.

the parser moves into its own module so the rule can be tested without pulling the server component and its fs work into the suite.
